Update base version of code

+++ CodingDojo/games/minesweeper/src/test/java/com/codenjoy/dojo/minesweeper/services/ScoresTest.java
@@ -4,7 +4,7 @@ package com.codenjoy.dojo.minesweeper.services;
  * #%L
  * Codenjoy - it's a dojo-like platform from developers to developers.
  * %%
- * Copyright (C) 2016 Codenjoy
+ * Copyright (C) 2018 Codenjoy
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -27,13 +27,8 @@ import com.codenjoy.dojo.services.PlayerScores;
 import com.codenjoy.dojo.services.settings.SettingsImpl;
 import org.junit.Test;
 
-import static junit.framework.Assert.assertEquals;
+import static org.junit.Assert.assertEquals;
 
-/**
- * User: sanja
- * Date: 05.06.13
- * Time: 20:35
- */
 public class ScoresTest {
     private PlayerScores scores;
     private SettingsImpl parameters = new SettingsImpl();
@@ -157,7 +152,7 @@ public class ScoresTest {
         minesweeperWin();    // +300
 
         Integer score = parameters.getParameter("Win score").type(Integer.class).getValue();
-        assertEquals(score.intValue(), scores.getScore());
+        assertEquals(score, scores.getScore());
     }
 
     @Test
@@ -167,7 +162,7 @@ public class ScoresTest {
         minesweeperClearBoard();    // +1
 
         Integer score = parameters.getParameter("Clear board score").type(Integer.class).getValue();
-        assertEquals(score.intValue(), scores.getScore());
+        assertEquals(score, scores.getScore());
 
     }
